[Refactor] Move Hive sink out of Exec
