+++ be/src/exec/CMakeLists.txt
@@ -291,7 +291,6 @@ set(EXEC_DATA_SINK_COMPAT_FILES
     data_sinks/data_sink_factory.cpp
     data_sinks/data_sink_pipeline.cpp
     data_sinks/pipeline_sink_builder.cpp
-    data_sinks/hive_table_sink.cpp
     data_sinks/hive_table_sink_pipeline_builder.cpp
     data_sinks/iceberg_table_sink.cpp
     data_sinks/table_function_table_sink.cpp
