+++ be/src/exec/data_sinks/data_sink_pipeline.cpp
@@ -20,9 +20,9 @@
 #include "data_sink/exchange/data_stream_sender.h"
 #include "data_sink/external/blackhole_table_sink.h"
 #include "data_sink/external/export_sink.h"
+#include "data_sink/external/hive_table_sink.h"
 #include "data_sink/tablet/multi_olap_table_sink.h"
 #include "data_sink/tablet/olap_table_sink.h"
-#include "exec/data_sinks/hive_table_sink.h"
 #include "exec/data_sinks/hive_table_sink_pipeline_builder.h"
 #include "exec/pipeline/exchange/exchange_sink_operator.h"
 #include "exec/pipeline/exchange/multi_cast_local_exchange.h"
